add layout page
