+++ views/layout/index.js
@@ -1,0 +1,12 @@
+import React, { Component, PropTypes } from 'react'
+
+export default class LayoutView extends Component {
+  render () {
+    const { main } = this.props
+    return (
+      <div className="container">
+        {main}
+      </div>
+    );
+  }
+}
